+++ GivingBackChange/GivingBackChange.Entity/Coin.cs
@@ -4,7 +4,7 @@ namespace GivingBackChange.Entity
 {
     public class Coin : EntityBase
     {
-        public Coin(int id, string label, float value, int quantity)
+        public Coin(int id, string label, decimal value, int quantity)
         {
             this.Id = id;
             this.Label = label;
@@ -14,7 +14,7 @@ namespace GivingBackChange.Entity
 
         public string Label { get; set; }
 
-        public float Value { get; set; }
+        public decimal Value { get; set; }
 
         public int Quantity { get; set; }
     }
